HuggingFace support for Quality Classifier (#190)
* update examples/scripts


* add hf support


* isort


* add ryan's review


---------
